# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Family Coupon Manager API
  description: |
    A secure, private web application API for managing family coupon codes. 
    This API provides endpoints for user authentication, coupon management, 
    search and filtering capabilities, and comprehensive audit logging.
    
    ## Authentication
    This API uses JWT (JSON Web Tokens) for authentication. After successful login,
    you'll receive an access token that must be included in the Authorization header
    for protected endpoints.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse:
    - Registration: 5 requests per 15 minutes per IP
    - Login: 10 requests per 15 minutes per IP
    - General API: 100 requests per 15 minutes per IP
    
    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.
    
  version: 1.0.0
  contact:
    name: Family Coupon Manager
    email: support@familycouponmanager.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Create a new user account with email and password. 
        Passwords are hashed using Argon2id for security.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              valid_registration:
                summary: Valid registration
                value:
                  email: "john.doe@example.com"
                  password: "SecurePass123!"
                  firstName: "John"
                  lastName: "Doe"
                  role: "MEMBER"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              examples:
                success:
                  summary: Successful registration
                  value:
                    success: true
                    data:
                      user:
                        id: "123e4567-e89b-12d3-a456-426614174000"
                        email: "john.doe@example.com"
                        firstName: "John"
                        lastName: "Doe"
                        role: "MEMBER"
                        isActive: true
                        createdAt: "2024-01-15T10:30:00Z"
                        updatedAt: "2024-01-15T10:30:00Z"
                      message: "User registered successfully"
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: |
        Authenticate user with email and password. 
        Returns access token and sets refresh token as httpOnly cookie.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              valid_login:
                summary: Valid login
                value:
                  email: "john.doe@example.com"
                  password: "SecurePass123!"
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: Refresh token set as httpOnly cookie
              schema:
                type: string
                example: "refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Strict; Max-Age=604800"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  summary: Successful login
                  value:
                    success: true
                    data:
                      user:
                        id: "123e4567-e89b-12d3-a456-426614174000"
                        email: "john.doe@example.com"
                        firstName: "John"
                        lastName: "Doe"
                        role: "MEMBER"
                        isActive: true
                        createdAt: "2024-01-15T10:30:00Z"
                        updatedAt: "2024-01-15T10:30:00Z"
                        lastLoginAt: "2024-01-15T14:30:00Z"
                      accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      message: "Login successful"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: |
        Logout the authenticated user and invalidate refresh token.
        Clears the refresh token cookie.
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                success:
                  summary: Successful logout
                  value:
                    success: true
                    data:
                      message: "Logout successful"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Refresh the access token using the refresh token stored in httpOnly cookie.
        Returns a new access token and refresh token.
      operationId: refreshToken
      responses:
        '200':
          description: Tokens refreshed successfully
          headers:
            Set-Cookie:
              description: New refresh token set as httpOnly cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
              examples:
                success:
                  summary: Successful token refresh
                  value:
                    success: true
                    data:
                      accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      message: "Tokens refreshed successfully"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get user profile
      description: Get the authenticated user's profile information
      operationId: getUserProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
              examples:
                success:
                  summary: User profile
                  value:
                    success: true
                    data:
                      user:
                        id: "123e4567-e89b-12d3-a456-426614174000"
                        email: "john.doe@example.com"
                        firstName: "John"
                        lastName: "Doe"
                        role: "MEMBER"
                        isActive: true
                        createdAt: "2024-01-15T10:30:00Z"
                        updatedAt: "2024-01-15T10:30:00Z"
                        lastLoginAt: "2024-01-15T14:30:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Coupon Management Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login endpoint. 
        Include in Authorization header as: `Bearer <token>`
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          description: User's email address
          example: "john.doe@example.com"
        firstName:
          type: string
          description: User's first name
          example: "John"
        lastName:
          type: string
          description: User's last name
          example: "Doe"
        role:
          $ref: '#/components/schemas/UserRole'
        isActive:
          type: boolean
          description: Whether the user account is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last account update timestamp
          example: "2024-01-15T10:30:00Z"
        lastLoginAt:
          type: string
          format: date-time
          description: Last login timestamp
          example: "2024-01-15T14:30:00Z"
          nullable: true
    UserRole:
      type: string
      enum: [ADMIN, MEMBER]
      description: User role in the system
      example: "MEMBER"
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address
          example: "john.doe@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]'
          description: |
            Password must contain at least:
            - 8 characters
            - One lowercase letter
            - One uppercase letter  
            - One number
            - One special character (@$!%*?&)
          example: "SecurePass123!"
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          description: User's first name
          example: "John"
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          description: User's last name
          example: "Doe"
        role:
          $ref: '#/components/schemas/UserRole'
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address
          example: "john.doe@example.com"
        password:
          type: string
          minLength: 1
          maxLength: 128
          description: User's password
          example: "SecurePass123!"

    # Coupon Schemas
...
    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            message:
              type: string
              example: "User registered successfully"
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            accessToken:
              type: string
              description: JWT access token (expires in 15 minutes)
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            message:
              type: string
              example: "Login successful"
    RefreshResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            accessToken:
              type: string
              description: New JWT access token
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            message:
              type: string
              example: "Tokens refreshed successfully"
    ProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
              example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Validation failed"
            details:
              description: Additional error details (optional)
              example: 
                field: "email"
                issue: "Invalid email format"
            timestamp:
              type: string
              format: date-time
              description: Error timestamp
              example: "2024-01-15T10:30:00Z"
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              summary: Validation error
              value:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Validation failed"
                  details:
                    field: "email"
                    issue: "Invalid email format"
                  timestamp: "2024-01-15T10:30:00Z"
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              summary: Unauthorized access
              value:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Authentication required"
                  timestamp: "2024-01-15T10:30:00Z"
            invalid_token:
              summary: Invalid token
              value:
                success: false
                error:
                  code: "INVALID_TOKEN"
                  message: "Invalid or expired token"
                  timestamp: "2024-01-15T10:30:00Z"
...
    ConflictError:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            conflict:
              summary: Resource conflict
              value:
                success: false
                error:
                  code: "CONFLICT"
                  message: "Resource already exists"
                  details: "Email address already registered"
                  timestamp: "2024-01-15T10:30:00Z"
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit:
              summary: Rate limit exceeded
              value:
                success: false
                error:
                  code: "RATE_LIMIT_EXCEEDED"
                  message: "Too many requests - please try again later"
                  details: "Rate limit: 10 requests per 15 minutes"
                  timestamp: "2024-01-15T10:30:00Z"
...